fix(tests): preserve proposal collection and project custody
