+++ backend/tests/projects/guide_compilation/proposals/test_contracts.py
@@ -5,49 +5,20 @@ from uuid import uuid4
 import pytest
 from pydantic import ValidationError
 
-from app.modules.projects.api.compilation_identity import CompilationComponentHashes
 from app.modules.projects.api.guide_proposals import (
     GuideProposalApproval,
     GuideProposalCorrection,
     GuideProposalTarget,
 )
 
-HASH = "sha256:" + "a" * 64
+from .contract_support import HASH, target_values
+
 POLICY_FIELDS = (
     "artifact_policy_id",
     "artifact_policy_hash",
     "artifact_projection_operation_id",
     "artifact_projection_output_digest",
 )
-
-
-def target_values():
-    return dict(
-        project_id=uuid4(),
-        guide_id=uuid4(),
-        compilation_id=uuid4(),
-        guide_version="1",
-        source_snapshot_id=uuid4(),
-        source_snapshot_hash=HASH,
-        setup_run_id=uuid4(),
-        setup_generation=1,
-        finalization_id=uuid4(),
-        finalization_facts_digest=HASH,
-        result_hash=HASH,
-        component_hashes={name: HASH for name in CompilationComponentHashes.model_fields},
-        pre_catalogue_id="pre",
-        pre_catalogue_version="1",
-        pre_catalogue_schema_version="1",
-        pre_catalogue_manifest_hash=HASH,
-        post_catalogue_id="post",
-        post_catalogue_version="1",
-        post_catalogue_schema_version="1",
-        post_catalogue_manifest_hash=HASH,
-        artifact_policy_id=uuid4(),
-        artifact_policy_hash=HASH,
-        artifact_projection_operation_id=uuid4(),
-        artifact_projection_output_digest=HASH,
-    )
 
 
 @pytest.mark.parametrize("field", POLICY_FIELDS)
